third_party/WebKit/LayoutTests/virtual/stable/fast/css3-text/css3-text-decoration/stable/getComputedStyle-text-decoration-expected.txt

Test that computed style for text-decoration is correctly represented.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS decoration is 'underline'
PASS successfullyParsed is true

TEST COMPLETE
This text should be underlined
